main/requirements.txt
# Core data analysis packages
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Interactive visualizations
plotly>=5.3.0

# API client
python-binance>=1.0.16

# Machine learning and statistics
scikit-learn>=1.0.0
xgboost>=1.5.0
scipy>=1.7.0
statsmodels>=0.13.0

# Rich console output
rich>=10.12.0

# Utilities
tqdm>=4.62.0
requests>=2.26.0 